ananya-flw/src/main/java/org/motechproject/ananya/service/OperatorService.java
package org.motechproject.ananya.service;

import org.motechproject.ananya.repository.AllOperators;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class OperatorService {

    private AllOperators allOperators;

    @Autowired
    public OperatorService(AllOperators allOperators) {
        this.allOperators = allOperators;
    }

    public Integer findMaximumUsageFor(String operator) {
        return allOperators.findByName(operator).getAllowedUsagePerMonth();
    }
}
